+++ src/main/scala/info/makingsense/sas/mapreduce/SasRecordReader.scala
@@ -1,4 +1,5 @@
 // Copyright (C) 2018 Forest Fang.
+// Modified 2026 by Making Sense.
 // See the LICENCE.txt file distributed with this work for additional
 // information regarding copyright ownership.
 //
@@ -14,11 +15,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.github.saurfang.sas.mapreduce
+package info.makingsense.sas.mapreduce
 
 import java.io.IOException
 
-import com.github.saurfang.sas.parso.ParsoWrapper
+import info.makingsense.sas.parso.ParsoWrapper
 import org.apache.commons.compress.compressors.bzip2.BZip2CompressorInputStream
 import org.apache.commons.io.input.CountingInputStream
 import org.apache.hadoop.fs.Path