+++ src/test/scala/info/makingsense/sas/SasInputFormatSpec.scala
@@ -1,4 +1,5 @@
 // Copyright (C) 2018 Forest Fang.
+// Modified 2026 by Making Sense.
 // See the LICENCE.txt file distributed with this work for additional
 // information regarding copyright ownership.
 //
@@ -14,9 +15,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.github.saurfang.sas
+package info.makingsense.sas
 
-import com.github.saurfang.sas.mapreduce.SasInputFormat
+import info.makingsense.sas.mapreduce.SasInputFormat
 import org.apache.hadoop.fs.{FileSystem, Path}
 import org.apache.spark.TestSqlContext
 import org.apache.hadoop.io.NullWritable